pyproject.toml
[project]
name = "gluellm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "any-llm-sdk[openai]>=1.5.0",
    "click>=8.3.1",
    "jinja2>=3.1.6",
    "pydantic>=2.12.5",
    "rich>=14.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
